Database.getAll(ofType:whereField:isEqualTo:completion:)
Added functionality to get all documents from RTD where a specific field is equal to a specific value. 

NOTE: Unfortunately, RTD does not work the same way Firestore does, so this method is MUCH less performant than it's Firestore counterpart and should be used as sparingly as possible. Should consider adding this note to the doc strings when we get around to adding doc strings to this project.